+++ profiles/jest.js
@@ -4,7 +4,7 @@ import jestFormatting from 'eslint-plugin-jest-formatting'
 
 import { base } from './test-base.js'
 
-import { apply, mergeRules } from '../dist/index.js'
+import { apply, merge, mergeRules } from '../dist/index.js'
 
 const appliedConfig = apply({
     jest,
@@ -15,10 +15,7 @@ const appliedConfig = apply({
 export default [
     {
         ...base,
-        plugins: {
-            ...base.plugins,
-            ...appliedConfig.plugins
-        },
+        plugins: merge(base.plugins, appliedConfig.plugins),
         rules: mergeRules(base, appliedConfig),
         settings: {
             jest: {