+++ src/index.ts
@@ -1,12 +1,8 @@
-export { default as apply } from './apply'
-export { default as applyPrettier } from './applyPrettier'
-export { default as ban } from './ban'
-export { default as deleteRules } from './deleteRules'
-export { default as mergeRules } from './mergeRules'
-export { default as replace } from './replace'
-
 export { default as merge } from './utils/merge'
+export { default as mergeArr } from './utils/mergeArr'
 
+export * from './tasks'
+export * from './plugins'
 export * from './constants'
 export * from './globs'
 export * from './lists'