+++ src/plugins/index.ts
@@ -1,0 +1,3 @@
+export { default as applyPrettier } from './applyPrettier'
+export { default as findTSConfigs } from './findTSConfigs'
+export { default as parseGitignore } from './parseGitIgnore'