+++ src/tasks/mergeRules.ts
@@ -1,6 +1,6 @@
-import type { Config, Rules } from './types'
-import isConfig from './utils/isConfig'
-import merge from './utils/merge'
+import type { Config, Rules } from '../types'
+import isConfig from '../utils/isConfig'
+import merge from '../utils/merge'
 
 export default function mergeRules(...rules: (Config | Rules)[]): Rules {
     const len = rules.length