+++ src/utils/mergeArr.ts
@@ -1,0 +1,4 @@
+export default function mergeArr<T, U>(target: T[], source: U[] | readonly U[]): T[] {
+    Array.prototype.push.apply(target, source as any)
+    return target
+}